[FEATURE-5] Move Link to Shared folder
